+++ bench.ts
@@ -1,32 +1,55 @@
 import * as denoRedis from "@db/redis";
 import { Redis } from "ioredis";
 import { createClient } from "redis";
-
-import { RedisClient } from "./mod.ts";
+import { RedisClient as DenoRedisClient } from "./deno.ts";
+import { RedisClient as WebRedisClient } from "./web.ts";
 
 const hostname = "127.0.0.1";
 const port = 6379;
 
 const redisConn = await Deno.connect({ hostname, port });
-const redisClient = new RedisClient(redisConn);
+const denoRedisClient = new DenoRedisClient(redisConn);
+const webRedisClient = new WebRedisClient(redisConn);
 const denoRedisConn = await denoRedis.connect({ hostname, port });
 const ioRedis = new Redis();
 
 const nodeRedisClient = await createClient().connect();
 
 Deno.bench({
-  name: "@iuioiua/redis",
+  name: "@iuioiua/redis (deno)",
+  ignore: true,
   baseline: true,
   async fn() {
-    await redisClient.sendCommand(["PING"]);
+    await denoRedisClient.sendCommand(["PING"]);
 
-    await redisClient.sendCommand(["SET", "mykey", "Hello"]);
-    await redisClient.sendCommand(["GET", "mykey"]);
+    await denoRedisClient.sendCommand(["SET", "mykey", "Hello"]);
+    await denoRedisClient.sendCommand(["GET", "mykey"]);
 
-    await redisClient.sendCommand(["HSET", "hash", "a", "foo", "b", "bar"]);
-    await redisClient.sendCommand(["HGETALL", "hash"]);
+    await denoRedisClient.sendCommand(["HSET", "hash", "a", "foo", "b", "bar"]);
+    await denoRedisClient.sendCommand(["HGETALL", "hash"]);
 
-    await redisClient.pipelineCommands([
+    await denoRedisClient.pipelineCommands([
+      ["INCR", "X"],
+      ["INCR", "X"],
+      ["INCR", "X"],
+      ["INCR", "X"],
+    ]);
+  },
+});
+
+Deno.bench({
+  name: "@iuioiua/redis (web)",
+  baseline: true,
+  async fn() {
+    await webRedisClient.sendCommand(["PING"]);
+
+    await webRedisClient.sendCommand(["SET", "mykey", "Hello"]);
+    await webRedisClient.sendCommand(["GET", "mykey"]);
+
+    await webRedisClient.sendCommand(["HSET", "hash", "a", "foo", "b", "bar"]);
+    await webRedisClient.sendCommand(["HGETALL", "hash"]);
+
+    await webRedisClient.pipelineCommands([
       ["INCR", "X"],
       ["INCR", "X"],
       ["INCR", "X"],
