+++ deno.json
@@ -1,7 +1,11 @@
 {
   "name": "@iuioiua/redis",
   "version": "1.0.1",
-  "exports": "./mod.ts",
+  "exports": {
+    "./": "./mod.ts",
+    "./deno": "./deno.ts",
+    "./web": "./web.ts"
+  },
   "imports": {
     "@db/redis": "jsr:@db/redis@^0.38.0",
     "@std/assert": "jsr:@std/assert@^1.0.11",
