+++ web_test.ts
@@ -3,7 +3,13 @@ import { Buffer } from "@std/streams/buffer";
 
 import { type Command, RedisClient, type Reply } from "./web.ts";
 
+const redisConn = await Deno.connect({ port: 6379 });
+const redisClient = new RedisClient(redisConn);
 const encoder = new TextEncoder();
+
+async function sendCommandTest(command: Command, expected: Reply) {
+  assertEquals(await redisClient.sendCommand(command), expected);
+}
 
 async function readReplyTest(output: string, expected: Reply, raw = false) {
   const redisClient = new RedisClient(new Buffer(encoder.encode(output)));
@@ -134,15 +140,6 @@ Deno.test("readReply() - large reply", async () => {
   await readReplyTest(`$${reply.length}\r\n${reply}\r\n`, reply);
 });
 
-const redisConn = await Deno.connect({ port: 6379 });
-const redisClient = new RedisClient(redisConn);
-
-await redisClient.sendCommand(["FLUSHALL"]);
-
-async function sendCommandTest(command: Command, expected: Reply) {
-  assertEquals(await redisClient.sendCommand(command), expected);
-}
-
 Deno.test("RedisClient.sendCommand() - transactions", async () => {
   await sendCommandTest(["MULTI"], "OK");
   await sendCommandTest(["INCR", "FOO"], "QUEUED");
@@ -212,11 +209,3 @@ Deno.test("RedisClient.writeCommand() + RedisClient.readReplies()", async () => 
     done: false,
   });
 });
-
-/* Deno.test("RedisClient.sendCommand() - no reply", async () => {
-  await assertRejects(
-    async () => await redisClient.sendCommand(["SHUTDOWN"]),
-    TypeError,
-    "No reply received",
-  );
-}); */
